Barra de progreso y cambiar edit y create a overview

+++ Agenda3/src/main/java/ch/makery/address/MainApp.java
@@ -142,7 +142,7 @@ public class MainApp extends Application {
             controller.setDialogStage(dialogStage);
             numPerson.setValue(personData.size());
             controller.setPerson(person);
-            controller.setBarrita(am.setProgreso(numPerson.intValue()));
+            controller.setBarrita(am.setProgreso(numPerson));
             // Show the dialog and wait until the user closes it
             dialogStage.showAndWait();
